Build o_voxel._C for CPU (macOS): missing pybind11 entry point, narrowing errors, dangling-pointer bug
o_voxel.io/o_voxel.serialize both unconditionally `from . import _C`, but _C had never been
built here -- setup.py's own CPU-build branch requires src/ext_cpu.cpp, which didn't exist in
this vendored checkout, even though the real CPU geometry implementations
(flexible_dual_grid.cpp, volumetic_attr.cpp, 775/872 lines) already exist complete. Wrote
ext_cpu.cpp: ext.cpp (the CUDA-era entry point) trimmed to only the CPU-available bindings.

Two more real, mechanical compile-error classes surfaced once ext_cpu.cpp existed, all in
vendored code presumably validated only against a different (Linux/gcc or nvcc) toolchain:
- 8 `-Wc++11-narrowing` errors (size_t -> long long/int in brace-init lists), fatal under Apple
  clang: filter_parent.cpp:83,154; svo.cpp:74,136; filter_neighbor.cpp:80,166;
  flexible_dual_grid.cpp:441,447,453. Fixed with explicit casts at each site.
- 2 `invalid suffix 'd' on floating constant` (1e-6d, 0.0d -- not valid C++, only removing the
  suffix was needed since the literal is already type double) in flexible_dual_grid.cpp:307,337.

z_order_encode_cpu/decode_cpu and hilbert_encode_cpu/decode_cpu are implemented in
serialize/api.cu and z_order.cu/hilbert.cu, but as CPU (__host__) functions living in the same
.cu files as CUDA-only __global__ kernels -- can't be compiled without nvcc, and the .h files
jointly declare both sides so can't even be included in a CPU build. Wrote
serialize/serialize_cpu.cpp: the CPU logic ported verbatim (bit-identical algorithm) with CUDA
decorators removed, added to setup.py's cpu_sources.

Also required: o-voxel/third_party/eigen submodule was declared in .gitmodules but never
initialized in this checkout (empty directory) -- `git submodule update --init` (no commit
needed, the parent repo's gitlink already pointed at the right commit).

Found and fixed a real bug of my own while verifying: my first serialize_cpu.cpp draft stored
raw data_ptr()s from `.contiguous()` temporaries in separate statements before use -- the
temporary is destroyed at the end of that statement, so for any non-contiguous input (e.g. the
strided column slices serialize.py's encode_seq/decode_seq always produce) the pointer was
already dangling before the loop ran. Caught via a real round-trip test (trivial all-zero input
decoded to garbage), not caught by the build succeeding. Fixed by holding the contiguous copies
as named locals spanning the whole function. Verified: z_order and hilbert round-trip correctly
now, including explicitly with a genuinely non-contiguous strided tensor view.

Also added `no_file` support to ObjaverseXL.py's foreach_instance -- dual_grid.py/voxelize_pbr.py
(TRELLIS.2's own, unmodified) call foreach_instance(..., no_file=True, ...) since those stages
read mesh_dumps/pbr_dumps by sha256 rather than from a raw downloaded file; the ported v1
foreach_instance had no such mode. Not yet run against the real pilot data (next step).

Note (not fixed here, out of scope): the same contiguous()-temporary-then-store-pointer pattern
exists in vendored convert/volumetic_attr.cpp:739-815 (texture/factor pointer arrays) -- didn't
reproduce a failure there and didn't want to rewrite vendor code speculatively, but worth
revisiting if a real correctness issue ever shows up in PBR-attribute voxelization.

+++ o-voxel/src/convert/flexible_dual_grid.cpp
@@ -304,7 +304,7 @@ void boundry_qef(
         // Calculate the QEF for the edge (boundary) defined by v0 and v1
         Eigen::Vector3d dir(v1.x() - v0.x(), v1.y() - v0.y(), v1.z() - v0.z());
         double segment_length = dir.norm();
-        if (segment_length < 1e-6d) continue; // Skip degenerate edges (zero-length)
+        if (segment_length < 1e-6) continue; // Skip degenerate edges (zero-length)
         dir.normalize();  // unit direction vector
 
         // Projection matrix orthogonal to the direction: I - d d^T
@@ -334,7 +334,7 @@ void boundry_qef(
 
         Eigen::Vector3d tMax, tDelta;
         for (int axis = 0; axis < 3; ++axis) {
-            if (dir[axis] == 0.0d) {
+            if (dir[axis] == 0.0) {
                 tMax[axis] = std::numeric_limits<double>::infinity();
                 tDelta[axis] = std::numeric_limits<double>::infinity();
             } else {
@@ -438,19 +438,19 @@ void face_from_dual_vertices(
 
         // xy-plane
         if (is_intersected[2] && neigh_indices[0] != kInvalidIndex && neigh_indices[1] != kInvalidIndex && neigh_indices[2] != kInvalidIndex) {
-            int4 quad_indices{i, neigh_indices[0], neigh_indices[2], neigh_indices[1]};
+            int4 quad_indices{(int)i, (int)neigh_indices[0], (int)neigh_indices[2], (int)neigh_indices[1]};
             auto tri_indices = quad_to_2tri(dual_vertices, quad_indices);
             face_indices.insert(face_indices.end(), tri_indices.begin(), tri_indices.end());
         }
         // yz-plane
         if (is_intersected[0] && neigh_indices[1] != kInvalidIndex && neigh_indices[3] != kInvalidIndex && neigh_indices[5] != kInvalidIndex) {
-            int4 quad_indices{i, neigh_indices[1], neigh_indices[5], neigh_indices[3]};
+            int4 quad_indices{(int)i, (int)neigh_indices[1], (int)neigh_indices[5], (int)neigh_indices[3]};
             auto tri_indices = quad_to_2tri(dual_vertices, quad_indices);
             face_indices.insert(face_indices.end(), tri_indices.begin(), tri_indices.end());
         }
         // xz-plane
         if (is_intersected[1] && neigh_indices[0] != kInvalidIndex && neigh_indices[3] != kInvalidIndex && neigh_indices[4] != kInvalidIndex) {
-            int4 quad_indices{i, neigh_indices[0], neigh_indices[4], neigh_indices[3]};
+            int4 quad_indices{(int)i, (int)neigh_indices[0], (int)neigh_indices[4], (int)neigh_indices[3]};
             auto tri_indices = quad_to_2tri(dual_vertices, quad_indices);
             face_indices.insert(face_indices.end(), tri_indices.begin(), tri_indices.end());
         }
